* tp/Texinfo/XS/main/targets.c, tp/Texinfo/XS/parsetexi/labels.c (set_labels_identifiers_target): move set_labels_identifiers_target to labels.c.
* tp/Texinfo/XS/main/targets.c: remove destroy_label_list.  Add
static for functions.

* tp/Texinfo/XS/parsetexi/api.c,
tp/Texinfo/XS/structuring_transfo/transformations.c: use size_t for
document descriptor more.

+++ tp/Texinfo/XS/parsetexi/api.c
@@ -20,6 +20,7 @@
 #include <dirent.h>
 #include <string.h>
 #include <errno.h>
+#include <stddef.h>
 
 #include "global_commands_types.h"
 #include "tree_types.h"
@@ -47,7 +48,7 @@
 #include "indices.h"
 #include "api.h"
 
-static int
+static size_t
 initialize_parsing (enum context root_ct)
 {
   parsed_document = new_document ();
@@ -163,7 +164,7 @@ parse_file_path (const char *input_file_path, char **result)
 int
 parse_file (const char *input_file_path, int *status)
 {
-  int document_descriptor = initialize_parsing (ct_base);
+  size_t document_descriptor = initialize_parsing (ct_base);
   GLOBAL_INFO *global_info;
   char *input_file_name_and_directory[2];
   int input_error;
@@ -211,7 +212,7 @@ parse_file (const char *input_file_path, int *status)
 int
 parse_text (const char *string, int line_nr)
 {
-  int document_descriptor = initialize_parsing (ct_base);
+  size_t document_descriptor = initialize_parsing (ct_base);
 
   input_push_text (strdup (string), line_nr, 0, 0);
   parse_texi_document ();
@@ -225,7 +226,7 @@ int
 parse_string (const char *string, int line_nr)
 {
   ELEMENT *root_elt;
-  int document_descriptor = initialize_parsing (ct_line);
+  size_t document_descriptor = initialize_parsing (ct_line);
 
   root_elt = new_element (ET_root_line);
 
@@ -238,7 +239,7 @@ parse_string (const char *string, int line_nr)
 int
 parse_piece (const char *string, int line_nr)
 {
-  int document_descriptor = initialize_parsing (ct_base);
+  size_t document_descriptor = initialize_parsing (ct_base);
   ELEMENT *before_node_section, *document_root;
 
   before_node_section = setup_document_root_and_before_node_section ();
